simplify rpc
