+++ block-producer/src/tasks/mod.rs
@@ -1,1 +1,3 @@
 pub mod txqueue;
+
+pub mod txqueuerpc;
